Work-around a JSpec bug that caused bad output
The bug, which caused bad output when using shared behaviors, and a test
failed, can be seen by provoking a failure in the file
`spec/unit/spec.shared-behaviors.js`. For instance, removing the name
property from the "Administrator" so that "User should have name" will
fail.

The work-around is not to create a shallow copy of the spec when copying them
from the suite that we share specs with, but to do a member by member copy.
As a simple hack, the "extend" function is used to do the copying.